requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
magic-filter==1.0.12
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.1
openai==2.15.0
packaging==25.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.2
pydantic==2.11.10
pydantic_core==2.33.2
Pyrogram==2.0.106
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==22.5
pytz==2024.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
Telethon==1.42.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
